INTERNSHIP TASK-2
Creating payment receipts is a pretty common task, be it an e-commerce website or any local store for that matter.
Here, you have to create our own transaction receipts just by using python. We would be using reportlab to generate the PDFs. Generally, it comes as a built-in package but sometimes it might not be present too. If it’s not present, then simply type the following in your terminal.